Critique Listing Page with real data
Adds the critique listing page with real data to the faculty role.
A faculty member can press the critique button on the attention component to see the page and it will pull up all of the students that have signed up for the event.

- Adds a prop to the CritiqueListComponent called currentEvent
- Adds a methed called getStudents that gets all student info for the signed up students
- Adds an on mounted call to getStudents

Still need to work on styling a little because the time format is not how I want it to look, but it is functional