Adds Email Functionality
This adds email functionality to the frontend.
This adds the email service that allows the frontend to access sending a custom email and sending out a notification.

This make the system send a notification to the Student when an instructor saves a critique for that student. This also adds the student's actual instrument to the critique stuff because I forgot last time.